update actions, added token check to private route
