+++ src/components/PrivateRoute.jsx
@@ -1,18 +1,39 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { Redirect, Route } from "react-router-dom";
 import { connect } from "react-redux"
+import { getSaved } from "../store/actions"
 
-const PrivateRoute = ({ component: Component, ...rest }) => (
+const PrivateRoute = ({ component: Component, ...rest }) => {
+  return(
   <Route
     {...rest}
     render={props =>
-      rest.dashboard.currentUser ? (
-        <Component {...props} />
+      window.localStorage.getItem('token') ? (
+        <TokenCheck {...props} Component={Component}/>
       ) : (
         <Redirect to="/login" />
       )
     }
   />
-);
+)
+};
 
-export default connect((state) => ({...state}), {})(PrivateRoute);
+const TokenCheckComponent = ({Component, ...props}) => {
+  console.log(props)
+  const { getSaved, user } = props;
+  useEffect( () => {
+    getSaved()
+  }, [ getSaved ])
+  if (user.tokenRejected) {
+    window.localStorage.deleteItem('token')
+    return <Redirect to="/login" />
+  }
+  console.log(user)
+  console.log(user.currentUser)
+  if ( user.currentUser&&!(user.tokenRejected)) return <Component {...props} />
+  else return <h1>Loading...</h1>
+}
+
+const TokenCheck = connect((state) => ({...state}), { getSaved })(TokenCheckComponent);
+
+export default PrivateRoute;
