+++ src/store/actions/dashboardActions.js
@@ -1,4 +1,4 @@
-import { axiosWithAuth } from "../../utils/axiosWithAuth";
+//import { axiosWithAuth } from "../../utils/axiosWithAuth";
 import axios from 'axios';
 
 export const DASHBOARD_LOADING_START = "DASHBOARD_LOADING_START";
@@ -11,36 +11,3 @@ export const getDashboard = songid => dispatch => {
         .catch(err => dispatch({ type: DASHBOARD_LOADING_FAILURE, payload: err }))
 }
 
-export const LOGIN_LOADING_START = "LOGIN_LOADING_START";
-export const LOGIN_LOADING_SUCCESS = "LOGIN_LOADING_SUCCESS";
-export const LOGIN_LOADING_FAILURE = "LOGIN_LOADING_FAILURE";
-export const login = creds => dispatch => {
-    dispatch({ type: LOGIN_LOADING_START });
-    axios.post(`https://spotify-song-suggester.herokuapp.com/login`, `grant_type=password&username=${creds.username}&password=${creds.password}`, {
-        headers: {
-            Authorization: `Basic ${btoa('lambda-client:lambda-secret')}`,
-            'Content-Type': 'application/x-www-form-urlencoded'
-        }
-    })
-        .then(res => {
-            dispatch({ type: LOGIN_LOADING_SUCCESS, payload: creds.username })
-            window.localStorage.setItem('token', JSON.stringify(res.data.access_token))
-        })
-        .catch(err => dispatch({ type: LOGIN_LOADING_FAILURE, payload: err }))
-}
-
-export const REGISTER_LOADING_START = "REGISTER_LOADING_START";
-export const REGISTER_LOADING_SUCCESS = "REGISTER_LOADING_SUCCESS";
-export const REGISTER_LOADING_FAILURE = "REGISTER_LOADING_FAILURE";
-export const register = creds => dispatch => {
-    dispatch({ type: LOGIN_LOADING_START });
-    axios.post(`https://spotify-song-suggester.herokuapp.com/createnewuser`, creds)
-        .then(res => dispatch({type: REGISTER_LOADING_SUCCESS, payload: res.data}))
-        .catch(err => dispatch({ type: REGISTER_LOADING_FAILURE, payload: err }))
-}
-
-export const LOGOUT_USER = "LOGOUT_USER";
-export const logout = () => {
-    window.localStorage.removeItem('token')
-    return {type: LOGOUT_USER};
-}
