+++ src/store/actions/index.js
@@ -1,3 +1,3 @@
-export { getDashboard, login, register, logout } from './dashboardActions'
-export { getSaved } from './savedActions'
+export { getDashboard } from './dashboardActions'
+export { getSaved, login, register, logout} from './userActions'
 export { getSearch, search } from './searchActions'
